Modified map method
Modified MAP method. Instead of altering the original list, map now does
something different. The convert function is responsible for allocation
of memmory required for the result of convert function and it stores the
the address of the allocated memmory into the destination given.

Later the destination given to addToList function to add that to the new
list.